step3/requirements.txt
aws-sam-cli==0.6.0
configparser==3.5.0
future==0.16.0
futures==3.2.0
google-api-python-client==1.7.4
python-jose==3.0.1
requests==2.19.1
step2==0.1
urllib3==1.23